Use LBR instead of drawf call-graph.
This option is viable on Haswell or later Intel platforms. It is much
lower overhead while profiling, generates more compact perf data, and is
much quicker to collapse and generate flame graphs.

The downsides is that it has limited stack depth, does work on major
cloud providers and some others listed at
https://lwn.net/Articles/617097/ and
http://www.brendangregg.com/perf.html