Fix building on Windows
Prior to this commit, it was not possible to build on Windows
because of signal_hook. This commit changes the way signal_hook is used
to only be imported when on Unix-like systems.

Additionally, during
https://github.com/ferrous-systems/flamegraph/commit/49b1ad87cfaab97fe39e6ba5b3fb199241cd74c7#diff-b4aea3e418ccdb71239b96952d9cddb6R132
a variable was mistakenly left unchanged for the Windows variant, which
was making the build fail.