Fix 'extraneous argument' error on macOS
Pass the workload command as a single argument, rather than multiple
arguments. The 'arg()` method calls dtrace like

  -c "./path/to/bin my_arg1 my_arg2"

On the other hand, 'args()` calls dtrace like

  -c ./path/to/bin my_ar1 my_arg2

Notice the lack of quotes. This leads to the "extraneous argument"
error, since anything after ./path/to/bin isn't used as the value
to the '-c' flag.

See "The second one" in #72.

flamegraph-rs 0.3.0
dtrace 1.15