fix #168: get metadata from correct crate when using manifest-path
In the implementation of #121, `--manifest-path` is forwarded to
`cargo build` so the correct crate gets built and profiled, however it
is not forwarded to `cargo metadata`.

This could impact target inference, but more importantly prevented
`--manifest-path` from being used in order to profile a crate without
being in any crate root e.g. if the crate being profile works off of
`$PWD`, `cargo-flamegraph` was not able to run if said `PWD` was not
in a crate context.