Improve 'perf script' failure error
Report an error when running 'perf script' on the perf.data file fails.
Currently when 'perf script' fails it just passes empty data on to
`inferno` which results in an unclear error about missing stack data.

Instead, this change checks the exit status of 'perf script' and
immediately exits if it failed.