Add support for profiling unit-test style benchmarks
This change extends the program with support for profiling unit test
benchmarks, aka libtest benchmarks. Support is accomplished in a vein
similar to that of unit tests: we introduce the --unit-bench argument
that behaves very much like the existing --unit-test does, but invokes
benchmarks instead of tests.
We cannot piggy-back on the existing --bench argument for the same
reason (presumably) that --unit-test is separate from --test: --bench
and --test require a target to be provided, but --unit-test and
--unit-bench can work without an argument or can be a pattern of tests
to match against, but not a target.

Closes: #275

Signed-off-by: Daniel Müller <deso@posteo.net>