Add unit-test-kind option to allow choosing bin or lib
When running with the unit-test option, there was no way to choose
between a bin or lib target, so the result was always an error if
there were both types for the same target name. To allow running
unit tests in that case, this new option unit-test-kind may be used
to specify the type of target that we want.